Feature/backend authentication role

+++ backend/utils/cloudinary-utils.js
@@ -5,7 +5,7 @@ import fs from 'fs'
 cloudinary.config({ 
     cloud_name: process.env.CLOUDINARY_NAME, 
     api_key: process.env.CLOUDINARY_API_KEY, 
-    api_secret: process.env.CLOUDINARY_API_SECRET // Click 'View API Keys' above to copy your API secret
+    api_secret: process.env.CLOUDINARY_API_SECRET 
 });
 
 const uploadFile = async (localFilePath) => {
@@ -14,19 +14,27 @@ const uploadFile = async (localFilePath) => {
         {
             return null;
         }
+
         const response = await cloudinary.uploader.upload(localFilePath,{
             resource_type : 'auto'
 
         })
         
-        console.log(`File upload on cloudainary src : ${response.url}`);
+        if(!(response.secure_url))
+        {   
+            console.error(`File not uploaded on cloudainary`);
+            fs.unlinkSync(localFilePath)
+            return null;
+        }
+        console.log(`File uploaded on cloudainary src : ${response.secure_url}`);
+        
         // Once file is uploaded we will delete from the localPath
         fs.unlinkSync(localFilePath)
-        return response;
+        return response.secure_url;
 
     } catch (error) {
-        // this delete the file from the localStorage 
-        fs.unlinkSync(localFilePath)
+        console.error(`Error uploading file: ${error.message}`);
+        fs.unlinkSync(localFilePath)    // Delete file from local temp path
         return null;
     }
 }
@@ -36,7 +44,7 @@ const deleteFromCloudinary = async (publicID)=>{
        const result = await cloudinary.uploader.destroy(publicID)
         console.log("Delete Successfully..");
     } catch (error) {
-        console.log("Error deleting the cloudinary files");
+        console.log("Error deleting the cloudinary file");
     }
 }
